Add missing files from v2.10.0 branch of official sprockets repo
The Gemfile, Rakefile and tests did not exist in the local version of
the gem on which this repo was based, so this commit adds them.
(Based on the README, this backport seems to have been based on the
v2.10.0 tag of the official sprokets repo, so those are the tests/files
I have used.)

There are four failing tests on this branch, but they are also failing
on the v2.10.0 tag of the offical sprockets repo.